cache npth

+++ pkgs/npth/fetch.sh
@@ -1,8 +1,11 @@
 #!/bin/sh
 rm -rf "$_DESTDIR" "$_SRCDIR"
-printf "Downloading source...\n"
-curl -L -# -o src.tar.bz2 https://gnupg.org/ftp/gcrypt/npth/npth-1.8.tar.bz2
+ver='1.8'
+if [ ! -f "$_DLCACHE/npth-$ver.tar.bz2" ] ||
+    [ "$(sha256sum "$_DLCACHE/npth-$ver.tar.bz2" | awk '{print $1}')" != "8bd24b4f23a3065d6e5b26e98aba9ce783ea4fd781069c1b35d149694e90ca3e" ]; then
+    printf "Downloading source...\n"
+    curl -L -# -o "$_DLCACHE/npth-$ver.tar.bz2" "https://gnupg.org/ftp/gcrypt/npth/npth-$ver.tar.bz2" || exit 1
+fi
 printf "Unpacking source...\n"
-tar -C "$_TMP" -xf src.tar.bz2
-rm src.tar.bz2
+tar -C "$_TMP" -xf "$_DLCACHE/npth-$ver.tar.bz2"
 mv "$_TMP"/npth-* "$_SRCDIR"
